Divided the program into separate classes, methods into separate smaller methods, optimized code
